+++ src/plugins/utils/utils-array.ts
@@ -118,3 +118,19 @@ export function sumNumberArray(array: number[]): number {
 export function maxOfNumbers(arr: number[]): number {
     return Math.max(...arr);
 }
+
+
+/**
+ * Appends the given documents to the given array.
+ * This will mutate the first given array.
+ * Mostly used as faster alternative to Array.concat()
+ * because .concat() is so slow.
+ * @link https://www.measurethat.net/Benchmarks/Show/4223/0/array-concat-vs-spread-operator-vs-push#latest_results_block
+ */
+export function appendToArray<T>(ar: T[], add: T[] | readonly T[]): void {
+    const amount = add.length;
+    for (let i = 0; i < amount; ++i) {
+        const element = add[i];
+        ar.push(element);
+    }
+}